refactor: 代码重构，区分接口与实现类。
