+++ src/main/java/com/example/aitourism/AiTourismApplication.java
@@ -2,8 +2,10 @@ package com.example.aitourism;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class AiTourismApplication {
 
 	public static void main(String[] args) {
